Aligned `Installation` framework with UCX project (#32)
This pull request introduces several changes to the `Installation`
framework in order to align it with the UCX project. The main changes
are the addition of new methods to the `Installation` class, as well as
modifications to the tests in `test_installer.py`.

In the `Installation` class, several new methods have been added:
`load_local`, `username`, `remove`, `workspace_link`,
`workspace_markdown_link`, and `_host`. The `load_local` method loads an
object of a specific type from a local file, while `username` returns
the username associated with the installation. The `remove` method
removes the installation folder from the workspace, and `workspace_link`
and `workspace_markdown_link` return links to files in the workspace.
The `_host` method returns the host URL for the workspace.

In the `test_installer.py` file, the tests for the `InstallState` class
have been updated to include a new test for the `from_installation`
method, which creates an `InstallState` object from an existing
`Installation` object. Additionally, the `test_state_corrupt` test has
been added to test the behavior of the `InstallState` class when the
installation is not found.

Overall, this pull request aims to improve the functionality and
robustness of the `Installation` framework by adding new methods and
updating the tests.

Relevant PRs:
- https://github.com/databrickslabs/ucx/pull/860