Added automated upgrade framework (#50)
This release introduces several improvements and additions to the
databricks.labs.blueprint module. It includes an automated upgrade
framework with a new `upgrades.py` file, providing mechanisms for
managing and applying upgrades to the product. The `ProductInfo` class
now includes a `checkout_root` method, a `as_semver` property, a
`wheels` property, and a `_infer_version_file` class method for better
handling of versioning and wheel building processes. A new exception,
`SingleSourceVersionError`, has been added to indicate that the version
file is not found in the given project. Additionally, test code
organization has been improved with new directories and files for
fixtures and unit tests, along with new test cases and functions for the
upgrades functionality. The `test_wheels.py` file has also been updated
with new functions to check the version of the Databricks SDK and handle
cases where the version marker is missing or does not contain the
`__version__` variable. These changes enhance the functionality,
maintainability, and adaptability of the system.