Release v0.3.1 (#61)
* Fixed the order of marshal to handle Dataclass with as_dict before
other types to avoid SerdeError
([#60](https://github.com/databrickslabs/blueprint/issues/60)). In this
release, we have addressed an issue that caused a SerdeError during the
installation.save operation with a Dataclass object. The error was due
to the order of evaluation in the _marshal_dataclass method. The order
has been updated to evaluate the `as_dict` method first if it exists in
the Dataclass, which resolves the SerdeError. To ensure the correctness
of the fix, we have added a new test_data_class function that tests the
save and load functionality with a Dataclass object. The test defines a
Policy Dataclass with an `as_dict` method that returns a dictionary
representation of the object and checks if the file is written correctly
and if the loaded object matches the original object. This change has
been thoroughly unit tested to ensure that it works as expected.