Release v0.4.2 (#75)
* Fixed edge cases when loading typing.Dict, typing.List and
typing.ClassVar
([#74](https://github.com/databrickslabs/blueprint/issues/74)). In this
release, we have implemented changes to improve the handling of edge
cases related to the Python `typing.Dict`, `typing.List`, and
`typing.ClassVar` during serialization and deserialization of
dataclasses and generic types. Specifically, we have modified the
`_marshal` and `_unmarshal` functions to check for the `__origin__`
attribute to determine whether the type is a `ClassVar` and skip it if
it is. The `_marshal_dataclass` and `_unmarshal_dataclass` functions now
check for the `__dataclass_fields__` attribute to ensure that only
dataclass fields are marshaled and unmarshaled. We have also added a new
unit test for loading a complex data class using the `MockInstallation`
class, which contains various attributes such as a string, a nested
dictionary, a list of `Policy` objects, and a dictionary mapping string
keys to `Policy` objects. This test case checks that the installation
object correctly serializes and deserializes the `ComplexClass` instance
to and from JSON format according to the specified attribute types,
including handling of the `typing.Dict`, `typing.List`, and
`typing.ClassVar` types. These changes improve the reliability and
robustness of our library in handling complex data types defined in the
`typing` module.
* `MockPrompts.extend()` now returns a copy
([#72](https://github.com/databrickslabs/blueprint/issues/72)). In the
latest release, the `extend()` method in the `MockPrompts` class of the
`tui.py` module has been enhanced. Previously, `extend()` would modify
the original `MockPrompts` object, which could lead to issues when
reusing the same object in multiple places within the same test, as its
state would be altered each time `extend()` was called. This has been
addressed by updating the `extend()` method to return a copy of the
`MockPrompts` object with the updated patterns and answers, instead of
modifying the original object. This change ensures that the original
`MockPrompts` object can be securely reused in multiple test scenarios
without unintended side effects, preserving the integrity of the
original state. Furthermore, additional tests have been incorporated to
verify the correct behavior of both the new and original prompts.