Release v0.4.4 (#83)
* If `Threads.strict()` raises just one error, don't wrap it with
`ManyError`
([#79](https://github.com/databrickslabs/blueprint/issues/79)). The
`strict` method in the `gather` function of the `parallel.py` module in
the `databricks/labs/blueprint` package has been updated to change the
way it handles errors. Previously, if any task in the `tasks` sequence
failed, the `strict` method would raise a `ManyError` exception
containing all the errors. With this change, if only one error occurs,
that error will be raised directly without being wrapped in a
`ManyError` exception. This simplifies error handling and avoids
unnecessary nesting of exceptions. Additionally, the `__tracebackhide__`
dunder variable has been added to the method to improve the readability
of tracebacks by hiding it from the user. This update aims to provide a
more streamlined and user-friendly experience for handling errors in
parallel processing tasks.