Release v0.6.3 (#111)
* fixed `Command.get_argument_type` bug with `UnionType`
([#110](https://github.com/databrickslabs/blueprint/issues/110)). In
this release, the `Command.get_argument_type` method has been updated to
include special handling for `UnionType`, resolving a bug that caused
the function to crash when encountering this type. The method now
returns the string representation of the annotation if the argument is a
`UnionType`, providing more accurate and reliable results. To facilitate
this, modifications were made using the `types` module. Additionally,
the `foo` function has a new optional argument `optional_arg` of type
`str`, with a default value of `None`. This argument is passed to the
`some` function in the assertion. The `Prompts` type has been added to
the `foo` function signature, and an assertion has been added to verify
if `prompts` is an instance of `Prompts`. Lastly, the default value of
the `address` argument has been changed from an empty string to
"default", and the same changes have been applied to the
`test_injects_prompts` test function.