Release v0.7.0 (#116)
* Added `databricks.labs.blueprint.paths.WorkspacePath` as
`pathlib.Path` equivalent
([#115](https://github.com/databrickslabs/blueprint/issues/115)). This
commit introduces the `databricks.labs.blueprint.paths.WorkspacePath`
library, providing Python-native `pathlib.Path`-like interfaces to
simplify working with Databricks Workspace paths. The library includes
`WorkspacePath` and `WorkspacePathDuringTest` classes offering advanced
functionality for handling user home folders, relative file paths,
browser URLs, and file manipulation methods such as `read/write_text()`,
`read/write_bytes()`, and `glob()`. This addition brings enhanced,
Pythonic ways to interact with Databricks Workspace paths, including
creating and moving files, managing directories, and generating
browser-accessible URIs. Additionally, the commit includes updates to
existing methods and introduces new fixtures for creating notebooks,
accompanied by extensive unit tests to ensure reliability and
functionality.
* Added propagation of `blueprint` version into `User-Agent` header when
it is used as library
([#114](https://github.com/databrickslabs/blueprint/issues/114)). A new
feature has been introduced in the library that allows for the
propagation of the `blueprint` version and the name of the command line
interface (CLI) command used in the `User-Agent` header when the library
is utilized as a library. This feature includes the addition of two new
pairs of `OtherInfo`: `blueprint/X.Y.Z` to indicate that the request is
made using the `blueprint` library and `cmd/<name>` to store the name of
the CLI command used for making the request. The implementation involves
using the `with_user_agent_extra` function from `databricks.sdk.config`
to set the user agent consistently with the Databricks CLI. Several
changes have been made to the test file for `test_useragent.py` to
include a new test case, `test_user_agent_is_propagated`, which checks
if the `blueprint` version and the name of the command are correctly
propagated to the `User-Agent` header. A context manager
`http_fixture_server` has been added that creates an HTTP server with a
custom handler, which extracts the `blueprint` version and the command
name from the `User-Agent` header and stores them in the `user_agent`
dictionary. The test case calls the `foo` command with a mocked
`WorkspaceClient` instance and sets the `DATABRICKS_HOST` and
`DATABRICKS_TOKEN` environment variables to test the propagation of the
`blueprint` version and the command name in the `User-Agent` header. The
test case then asserts that the `blueprint` version and the name of the
command are present and correctly set in the `user_agent` dictionary.
* Bump actions/checkout from 4.1.6 to 4.1.7
([#112](https://github.com/databrickslabs/blueprint/issues/112)). In
this release, the version of the "actions/checkout" action used in the
`Checkout Code` step of the acceptance workflow has been updated from
4.1.6 to 4.1.7. This update may include bug fixes, performance
improvements, and new features, although specific changes are not
mentioned in the commit message. The `Unshallow` step remains unchanged,
continuing to fetch and clean up the repository's history. This update
ensures that the latest enhancements from the "actions/checkout" action
are utilized, aiming to improve the reliability and performance of the
code checkout process in the GitHub Actions workflow. Software engineers
should be aware of this update and its potential impact on their
workflows.

Dependency updates:

* Bump actions/checkout from 4.1.6 to 4.1.7
([#112](https://github.com/databrickslabs/blueprint/pull/112)).