`WorkspacePath`: Fix the .rename() and .replace() implementations to return the target path (#130)
This PR updates the `.rename()` and `.replace()` implementations on
`WorkspacePath` so that they return the target path. (This is part of
the public API.)

In addition, the `.rename()` method no longer extends the public API
with the additional `overwrite` keyword argument. The contract for
`.rename()` is that it is allowed to always fail if the target exists,
which is how it now behaves.

Out of scope for this PR is defining the exact error that is raised if
the target exists.

This PR is part of the preparation for factoring out common code to
implement support for DBFS paths.