Release v0.8.2 (#139)
* Make hatch a prerequisite
([#137](https://github.com/databrickslabs/blueprint/issues/137)). In
version 1.9.4, hatch has become a prerequisite for installation in the
GitHub workflow for the project's main branch, due to occasional
failures in `pip install hatch` that depend on the local environment.
This change, which includes defining the hatch version as an environment
variable and adding a new step for installing hatch with a specific
version, aims to enhance the reliability of the build and testing
process by eliminating potential installation issues with hatch. Users
should install hatch manually before executing the Makefile, as the line
`pip install hatch` has been removed from the Makefile. This change
aligns with the approach taken for ucx, and users are expected to
understand the requirement to install prerequisites before executing the
Makefile. To contribute to this project, please install hatch using `pip
install hatch`, clone the GitHub repository, and run `make dev` to start
the development environment and install necessary dependencies.
* support files with unicode BOM
([#138](https://github.com/databrickslabs/blueprint/issues/138)). The
recent change to the open-source library introduces support for handling
files with a Unicode Byte Order Mark (BOM) during file upload and
download operations in Databricks Workspace. This new functionality,
added to the `WorkspacePath` class, allows for easier reading of text
from files with the addition of a `read_text` method. When downloading a
file, if it starts with a BOM, it will be detected and used for
decoding, regardless of the preferred encoding based on the system's
locale. The change includes a new test function that verifies the
accurate encoding and decoding of files with different types of BOM
using the appropriate encoding. Despite the inability to test Databrick
notebooks with a BOM due to the Databricks platform modifying the
uploaded data, this change enhances support for handling files with
various encodings and BOM, improving compatibility with a broader range
of file formats, and ensuring more accurate handling of files with BOM.