Release v0.10.0 (#184)
* Fixed incorrect script for no-pylint-disable
([#178](https://github.com/databrickslabs/blueprint/issues/178)). In
this release, we have updated the script used in the `no-cheat` GitHub
workflow to address false positives in stacked pull requests. The
updated script fetches the base reference from the remote repository and
generates a diff between the base reference and the current branch,
saving it to a file. It then runs the "no_cheat.py" script against this
diff file and saves the results to a separate file. If the count of
cheats (instances where linting has been intentionally disabled) is
greater than one, the script outputs the contents of the results file
and exits with a non-zero status, indicating an error. This change
enhances the accuracy of the script and ensures it functions correctly
in a stacked pull request scenario. The `no_cheat` function, which
checks for the presence of certain pylint disable tags in a given diff
text, has been updated to the latest version from the ucx project to
improve accuracy. The function identifies tags by looking for lines
starting with `-` or "+" followed by the disable tag and a list of
codes, and counts the number of times each code is added and removed,
reporting any net additions.
* Skip dataclassess fields only when `None`
([#180](https://github.com/databrickslabs/blueprint/issues/180)). In
this release, we have implemented a change that allows for the skipping
of dataclass fields only when the value is `None`, enabling the
inclusion of empty lists, strings, or zeros during marshalling. This
modification is in response to issue
[#179](https://github.com/databrickslabs/blueprint/issues/179) and
involves adding a check for `None` before marshalling a dataclass field.
Specifically, the previous condition `if not raw:` has been replaced
with `if raw is None:`. This change ensures that empty values such as
`[]`, `''`, or `0` are not skipped during the serialization process,
unless they are explicitly set to `None`. This enhancement provides
improved compatibility and flexibility for users working with
dataclasses containing empty values, allowing for more fine-grained
control during the serialization process.

Dependency updates:

* Bump codecov/codecov-action from 4 to 5
([#174](https://github.com/databrickslabs/blueprint/pull/174)).