Release v0.10.1 (#186)
* patch hosted runner
([#185](https://github.com/databrickslabs/blueprint/issues/185)). In
this release, we have implemented a temporary fix to address issues with
publishing artifacts in the release workflow. This fix involves changing
the runner used for the job from `ubuntu-latest` to a protected runner
group labeled "linux-ubuntu-latest". This ensures that the job runs on a
designated hosted runner with the specified configuration, enhancing the
reliability and security of the release process. The `permissions`
section of the job remains unchanged, allowing authentication to PyPI
and signing of release artifacts with sigstore-python. It is worth
noting that this is a stopgap measure, and further changes to the
release workflow may be made in the future.