Test coverage for the customised logger setup (#232)
This PR provides test coverage of the custom logger that blueprint
installs, and establishes the baseline behaviour. Changes include:

 - Type-hints for the logger module, and updated docstrings.
- Plumbing to support testing of the logger configuration that we
provide.
- A tweak to the stream that is probed, if we're going to probe: it
should be stderr, not stdout. (Currently the probing facility is unused,
I believe). Closes #227.
- Unit tests for the `install_logger()` method to verify the
configuration that it sets up.
 - Unit tests for the custom formatter: `NiceFormatter`
Some of the tests currently fail due to problems with the existing
implementation that will be fixed in subsequent PRs.

All updates to the external interfaces are API-compatible.

Relates: #229, #230.