Fix argument interpolation in colorised logs (#233)
This PR fixes an issue with the colorised log formatter whereby it
wasn't properly formatting log entries if they contained `%`-style
placeholders (with arguments). Although we don't generally use this
style in our own projects, 3rd-party code does and if we don't support
this then logging from those components is incorrect.

This is stacked on top of #232 and resolves #230.