Fix logger name abbreviation fails if the logger name contains `..` (#236)
When colorising logger output, the logger name (conventionally the
module) is abbreviated: this abbreviating throws an exception during
formatting if the logger name contains `..`. This is unlikely to occur
during normal use, but may occur if the logger is used in a script that
was referred to via a path that includes `..`.

This PR is stacked on top of #232.