Ensure the names of logger levels are consistent (#234)
This PR updates the names we use for the `WARNING` and `CRITICAL`
log-levels when produced colorised logs. Prior to this PR:

- When not-colorised, we used the conventional `WARNING` and `CRITICAL`
names.
 - When colorised, we used `WARN` and `FATAL` instead.

Although the latter is more compact, it's more important to: a) be
consistent; b) conform to the norms of the Python ecosystem and use the
conventional names for the logging levels.

This PR is stacked on top of #232 and resolves #229.