Consistent exception formatting in logs (#237)
This PR updates the format of logs to be consistent when exception
information is included, irrespective of whether the logs are colorised
or not. Prior to this PR, colorised logs were formatted slightly
differently to the Python norms:
```
16:06:10 ERROR [d.l.module] An error occurred: Traceback (most recent call last):
  File "/path/to/file.py", line 410, in some_function
    raise RuntimeError("Test exception.")
RuntimeError: Test exception.
```
(Note that the traceback starts on the line of the error.)

This PR updates this to be:
```
16:08:18 ERROR [d.l.module] An error occurred
Traceback (most recent call last):
  File "/path/to/file.py", line 410, in some_function
    raise RuntimeError("Test exception.")
RuntimeError: Test exception.
```

This PR is stacked on top of #232.