Marshalling: allow JSON-like fields (#241)
This PR continues from #189 by updating it with the following changes:

- It introduces a specific `JsonValue` type alias that represents the
maximum bounds of the values that can be saved for an installation.
- It removes support for `Any` and `object` as type annotations on data
classes that we marshal. The existing tests for these have been updated
to use `JsonValue` instead.
- Support for marshalling raw `list` and `dict` fields is deprecated.
Saving these used to work, but loading would fail. I've chosen to deal
with this by:
- Issuing a DeprecationWarning if they're hit when saving, with
instructions to use `list[T]` or `dict[T]` instead.
- During loading, a specific error is thrown indicating how it can be
fixed.
- Documents the restrictions and limitations on `.save()` for what it
can actually handle.
- Adds a marker for an improvement whereby during marshalling the type
hints should actually be checked against the values that are being
saved.
 - Fixed some tests that were accidentally passing.
- Tightens some of the type-hints we had in place where we declared
`Any` but really only accepted a subset.

This is sufficient to allow the use-case that we need in
databrickslabs/remorph#1488, and is preferred to the approach in #189
because it forces developers to fix their code:

- Declaring `JSONValue` as a field will let the type-checker tell them
if a value is assigned that does not conform. (The fix is for the client
code to validate the values.)
- Similarly during loading the type-checker forces the caller to check
and handle the possible types that might be present.

One of the bugs fixed in #189 relates to automatic conversion between
JSON values: for example, from `123` to `"123"` or vice-versa. Some
tests were (incorrectly) relying on this, as were some tests in a
downstream project (remorph). _There might be other places in production
code where the previous behaviour is relied upon._ As such, the next
release that includes this code (or that from #189) should bump the
minor version and not just be a patch release.

---------

Co-authored-by: Eric Vergnaud <eric.vergnaud@databricks.com>