Expose the number of available CPUs for concurrent processing (#244)
Blueprint already has a framework for concurrent running of tasks, with
the level of concurrency normally initialised based on the number of
CPUs the local system has. This PR:

 - Improves the way in which the CPU count is determined.
 - Exposes the CPU count for downstream use.

On modern systems, sometimes the number of CPUs available for use is
different to the number of CPUs present: this is especially true in
containerised environments, although CPU quotas can be imposed outside
of containers as well. When this is in place, the available CPU quota
should be used instead. This PR tries to determine this in various ways
before falling back on the number of CPUs present.

Downstream in Remorph we want configure some concurrent processing, and
given that it's mildly complex to determine the CPU count this PR
exposes that count so we can use it there.