Improve support for reading text files that contain a Unicode BOM at the start (#243)
This PR backports code from the UCX project for reading text files that
have a Unicode BOM at the start. (Python does not handle this out-of-box
when opening files in text mode.)

Blueprint already had support for handling this when open Workspace
files, but there wasn't support for doing this for local files as well.
As such this PR:

 - Includes the code that had in UCX for handling this situation;
 - Improves the support for non-seekable files.
- Refactors the existing code for Workspace files to use the same
implementation.

Tests:

 - Added unit tests.
 - Existing integration tests.

This PR is a pre-requisitie for databrickslabs/remorph#1516.