Support detecting file encoding from XML declaration (#254)
Blueprint has support for reading a text file and detecting its encoding
via a BOM-marker. This PR extends this for XML files, where the encoding
is specified in the declaration at the start of the file, for example:
```xml
<?xml version="1.0" encoding="iso-8859-1"?>
```

One thing to bear in mind for XML: the default encoding is UTF-8, unless
a BOM includes something else.