Fix configuration file unmarshalling of JSON floating-point values (#253)
This PR fixes unmarshalling of `JsonValue` fields if they contain
floating point values; prior to this PR they were incorrectly
unmarshalled as integers, truncating precision.

The reason for this error is that unmarshalling is somewhat forgiving
with coercion allowed between primitives: if the target is a string,
boolean/int/float will be converted to a string. Previous this allowed
`int(float_value)` to succeed; this PR blocks that specifically, and
unit tests have been updated to ensure the values being produced are of
the correct type.