Sniff encoding properly in XML files with a standalone directive (#256)
This PR updates the code that detects the encoding of XML files so that
it correctly detects the `encoding=` attribute if there is also a
`standalone=` directive. Previously we only supported XML directives of
the form:

```xml
<?xml version="1.x" encoding="xxx"?>
```

However the XML declaration may also contain a `standalone=` attribute.
If present we previously failed to detect the encoding attribute, but
now we also support XML directives of the form:
```xml
<?xml version="1.x" encoding="xxx" standalone="yes"?>
```