Improve `Prompts` type hints (#267)
This PR updates the type hints on the `Prompts` interface:

- No more `Any` (which is never checked), instead generic type arguments
are used.
- The interface for `choice()` now specifies that the choices must be
strings; the internal implementation assumed this anyway.
- Read-only (collection) types are used instead of `dict[]` and
`list[]`, indicating the arguments are not modified.

In addition, we now use (and allow) precompiled regexes for the
`valid_regex` argument on `Prompt.question()`.

These changes are backwards compatible at runtime, and mostly backwards
compatible for type-checking. The exception is the change to
`Prompt.choice()`: if the `choices` argument was not known to be a
sequence of strings, a type checker must now be able to infer this.